chore: speedup `npm install` from a github checkout
Many of us are running a watchdog in the background that
keeps `//lib` folder up-to-date, so there's no need to re-build
Playwright from source on every `npm install` command.

This patch does a crude timestatmp comparison between typescript
files and their javascript counterparts. If some files are missing in
`//lib`, or if some `//src` files are newer than `//lib`, then
it re-builds the Playwright.

This cuts off 10 seconds for me on in case I have a watchdog in the
background.