Add tests for Blob downloads (#1936)
This patch adds test cases for Blob downloads which appear to have
differing behavior on the different browsers.

NB: For each of the browsers, I manually tried going to a page that
looked like the newly added test asset page (`download-blob.html`)
and performing the steps that the test does. In each of the cases
I ended up with the target file; however, Firefox and Safari both had an
additional prompt. (I'm not sure if that prompt was a default prompt
that should be disabled when running via Playwright or if additional
handling logic is needed.)

For Chrome, this test passes (and manual replication yields a file
download).

For Firefox, this test crashes the browser (but manual replication
yields a file but requires dealing with one extra download prompt in
Firefox).

For Webkit, this test will timeout (but manual replication yields a file
but requires dealing with one extra download prompt in Webkit.)

Test Output:

```
Failures:
1) [CRASHED] Firefox Download should report download path within page.on('download', …) handler for Blobs (download.spec.js:86:3)

2) [TIMEOUT 10000ms] WebKit Download should report download path within page.on('download', …) handler for Blobs (download.spec.js:86:3)
```