chore(build): Allow building of package without Chromium
In order to generate TS types for this package, Chromium is required.
The usage of Chromium is tedious, especially when building Playwright
inside a Docker container. The main reason for problems is the sandbox
feature of Chromium, which has to be turned off in certain environments.
Additionally, Playwright types are not needed in a CI environment,
so building the types in a CI container is unnecessary.

By splitting the build command into two npm-scripts, a Playwright
build inside a Docker container can be done without having to run
Chromium.

Before:
* `node ./utils/runWebpack.js --mode='development'`
* `npx --no-install tsc -p .`
* `mkdir types`
* `./packages/build_package.js playwright playwright.tar.gz`

After:
* `npm run build-lib`
* `./packages/build_package.js playwright playwright.tar.gz`