devops: encode build number together with Chromium revision
This is an alternative approach to #3698 that was setting up a custom
mapping between chromium revisions and our mirrored builds.

One big downside of the opaque mapping was inability to quickly
understand which Chromium is mirrored to CDN.

To solve this, this patch starts treating browser revision as a fractional number,
with and integer part being a chromium revision, and fractional
part being our build number.

For example:
- current `BUILD_NUMBER` for Chromium that we mirror to our CDN is `792639`.
- if I want to change the packaging procedure, e.g. update bundled
FFMPEG, I can change `BUILD_NUMBER` to `792639.1`.

In the Playwright code itself, there are a handful of places that treat
browser revision as integer, mostly to compare them with some particular
revision numbers. This code would still work, but I changed these places
to parse float instead for correctness.