browser(webkit): change scrollIntoView to only scroll if needed
We used to have alignCenterAlways to workaround the RenderLayer logic
that does not scroll horizontally if enough of the element is visible.

This change plumbs the bit "avoid minimum threshold" in the
ScrollAlignment. This way, we don't actually scroll when the desired
rect is already visible and prevent overlay scrollbars flashing that
block pointer actions on MacOS.