feat(fill): make fill work when targeting elements inside the label
We already accept `<label>` elements as a target for `fill()`.
In this case, we fill the `label.control` element - that's usually
an `<input>` element.

However, there is often some markup inside the label, e.g.
`<span>` that makes part of the text bold. We do not support filling
using such an element currently, while we should be able to.