test(download): introducing a large payload for the download-delay test
It is observed that Chromium requires a large payload to trigger the
download event early enough, otherwise that might be delayed for several
seconds.
Before the change the "should be able to cancel pending downloads" test
case fails occasionally (e.g. 7 times out of 100 repeats). After the
change a local run of 1000 repeats suceeded with no failure.