feat(actions): ensure we perform the action on the expected element
This changes previous `checkHitTarget` heurisitc that took place
before the action, to `setupHitTargetInterceptor` that works during
the action:
- Before the action we set up capturing listeners on the window.
- During the action we ensure that event target is the element
  we expect to interact with.
- After the action we clear the listeners.

This should catch the "layout shift" issues where things move
between action point calculation and the actual action.
There is still a possiblity of mistargeting with iframes shifting
around, because we only intercept in the target frame.