fix(firefox): race between markAsError and Page.ready
When new page turns into download, we can have `_markAsError` followed
by the `Page.ready` signal and report the page twice (as an error and
as a real page). This is flaky and depends on whether `Page.ready` is
fast enough before the page closes or not.

Exposed by the "should report new window downloads" test.