browser(webkit): fix build with shallow checkouts
https://github.com/WebKit/WebKit/commit/79d0e78bb1237a3a2727a20faf10a37de601172c

The above commit does not account for the git repository possibly
being a shallow clone with modified HEAD - adding a local commit
on top of a clone like this results in the origin reference being
erased, which results in something like this:

fatal: ambiguous argument 'origin/HEAD': unknown revision or path not in the working tree.

Since we have no reference point there, let's just give up and
pretend as if the repository wasn't a git clone at all.