feat(launchServer): accept wsPath option
This change allows whoever is calling the launcher to pass
in a secret path to use, instead of having the PW code pick
for you.

This is particularly helpful in use cases where you have
a trusted process launching Docker containers containing the browser
server and you want the host launching the container to know what path
and port to access the running service at without grepping through logs
or resorting to other message passing mechanisms.

[Slack Thread][slack]

Closes #6291

[slack]: https://playwright.slack.com/archives/CSUHZPVLM/p1629509151068800?thread_ts=1629492900.067500&cid=CSUHZPVLM