feat(har): add _requestref
This will be used by our traceviewer-like (userland) debug tools
that require correlating logging emitted from Playwright-dependent code
and the requests that end up in the HAR file after the run.

This can be utilized by Playwright's own traceviewer some day, too. For
example, it can be used to link back code that transformed a request to
the actual request that ended up getting recorded in the HAR file.